Revert "Convert metadata files to current format on attach to dialog. Fix #378"
