+++ src/routers/Search/dashboard/src/components/TagSelector.svelte
@@ -3,8 +3,6 @@
   A component for selecting tags for a repository or content.
 -->
 <script lang="ts">
-  import type { ArtifactMetadatav2 } from '../../../src/adapters/common/types';
-  import { toArtifactMetadatav2 } from '../../../src/Utils';
   import type { default as ContentType } from "../ContentType";
   import Dropzone from "svelte-file-dropzone";
   import {
@@ -12,14 +10,14 @@
     readFile,
   } from "../Utils";
 
-  export let metadata: ArtifactMetadatav2;
+  export let metadata;
   export let contentType: ContentType;
   export let displayTypeName: string = 'content';
   export let disabled: boolean = false;
 
-  function getTagDisplayName(tags) {
+  function getTagDisplayName(tag) {
     // FIXME: there is no way to tell the difference btwn terms and compound fields...
-    let currentTag = tags;
+    let currentTag = tag;
     const tagNames = [];
     while (currentTag) {
       const [name, tag] =
@@ -37,7 +35,8 @@
     const [tagsFile] = event.detail.acceptedFiles;
     console.log('file dropped', tagsFile);
     if (tagsFile) {
-      metadata = toArtifactMetadatav2(JSON.parse(await readFile(tagsFile)));
+      metadata = JSON.parse(await readFile(tagsFile));
+      console.log({metadata})
     }
   }
 
@@ -46,7 +45,7 @@
 <div>
     <p> <!-- TODO: Check if they are actually optional -->
       Taxonomy Terms <span style="font-style:italic">(optional)</span>:<br />
-      {metadata && metadata.tags ? getTagDisplayName(metadata.tags).join(", ") : ""}
+      {metadata && metadata.taxonomyTags ? metadata.taxonomyTags.map(getTagDisplayName).join(", ") : ""}
     </p>
 
     {#if disabled}
