+++ test/routers/Search/Utils.spec.js
@@ -1,6 +1,6 @@
 describe("Utils", function () {
   const assert = require("assert");
-  const { retry, range, intervals, Pattern, toArtifactMetadatav2 } = require(
+  const { retry, range, intervals, Pattern } = require(
     "../../../src/routers/Search/build/Utils",
   );
 
@@ -132,79 +132,6 @@ describe("Utils", function () {
         assert(!validate("AAA"));
         assert(!validate("BBB"));
       });
-    });
-  });
-
-  describe("toArtifactMetadatav2", function () {
-    const time = new Date().toString();
-    const taxonomyVersion = {
-      id: "taxID",
-      commit: "abcdef",
-    };
-
-    it("should convert single tax tag correctly", function () {
-      const metadata = {
-        displayName: "Example",
-        taxonomyVersion,
-        time,
-        taxonomyTags: [
-          {
-            Base: {
-              name: {
-                value: "Hi there!",
-              },
-            },
-          },
-        ],
-      };
-      const updated = toArtifactMetadatav2(metadata);
-
-      assert(updated.tags);
-      assert.deepEqual(updated.tags, metadata.taxonomyTags[0]);
-    });
-
-    it("should initialize tags to {} if no taxonomyTags", function () {
-      const metadata = {
-        displayName: "Example",
-        taxonomyVersion,
-        time,
-      };
-      const updated = toArtifactMetadatav2(metadata);
-
-      assert(typeof updated.tags === "object");
-      assert.equal(Object.keys(updated.tags), 0);
-    });
-
-    it("should convert multi tax tags correctly", function () {
-      const metadata = {
-        displayName: "Example",
-        taxonomyVersion,
-        time,
-        taxonomyTags: [
-          {
-            Base: {
-              name: {
-                value: "Hi there!",
-              },
-            },
-          },
-          {
-            Base: {
-              content: {
-                type: "SomeType",
-                id: "/a/b/c",
-              },
-            },
-          },
-        ],
-      };
-      const updated = toArtifactMetadatav2(metadata);
-
-      assert(updated.tags);
-
-      assert.deepEqual(updated.tags.Base.name.value, "Hi there!");
-      assert.deepEqual(updated.tags.Base.content.type, "SomeType");
-      assert.deepEqual(updated.tags.Base.content.id, "/a/b/c");
     });
   });
 });
